requirements.txt
numpy==1.24.4
oauthlib==3.1.0
onnxruntime-1.11.1-cp39-cp39-linux_armv7l.whl
protobuf==6.30.2
PyAudio==0.2.11
sounddevice==0.5.1
soundfile==0.13.1
